remove more completely unused
